[MP-005] Moved pagination-related classes to a separate package, added @Profile.
