+++ scheduler-quartz/src/main/kotlin/io/scheduler/comparison/quartz/jobs/handlers/pagination/LocaLolaDedicatedJobHandler.kt
@@ -6,7 +6,7 @@ import io.scheduler.comparison.quartz.jobs.pagination.impl.listJobPaginator
 import io.scheduler.comparison.quartz.jobs.state.DedicatedOrderJobData
 import io.scheduler.comparison.quartz.jobs.state.DedicatedOrderJobMetadata
 import io.scheduler.comparison.quartz.messaging.LocaLolaRefundsSender
-import io.scheduler.comparison.quartz.repositories.LocaLolaFailuresRepository
+import io.scheduler.comparison.quartz.repositories.pagination.LocaLolaFailuresRepository
 import org.springframework.context.annotation.Profile
 import org.springframework.stereotype.Component
 import org.springframework.transaction.annotation.Transactional
