+++ scheduler-quartz/src/main/kotlin/io/scheduler/comparison/quartz/repositories/pagination/CommonOperationOnOrderRepository.kt
@@ -1,9 +1,10 @@
-package io.scheduler.comparison.quartz.repositories
+package io.scheduler.comparison.quartz.repositories.pagination
 
 import io.scheduler.comparison.quartz.domain.OperationOnOrder
 import io.scheduler.comparison.quartz.domain.OrderOperationStatus
-import io.scheduler.comparison.quartz.jobs.state.DedicatedOrderJobData
+import io.scheduler.comparison.quartz.jobs.state.CommonOrderJobData
 import org.intellij.lang.annotations.Language
+import org.springframework.context.annotation.Profile
 import org.springframework.jdbc.core.simple.JdbcClient
 import org.springframework.stereotype.Repository
 import org.springframework.transaction.annotation.Propagation
@@ -11,27 +12,26 @@ import org.springframework.transaction.annotation.Transactional
 import java.sql.Types
 import java.time.LocalDateTime
 
-// Todo: queryForStream with Stream item processing (pageSize == fetchSize), maxCount == LIMIT.
 @Repository
-class WildFruitOperationOnOrderRepository(
+@Profile("pagination")
+class CommonOperationOnOrderRepository(
     private val jdbcClient: JdbcClient
 ) {
 
     private companion object {
         @Language("PostgreSQL")
-        const val READ_UNPROCESSED_ORDER_OPERATIONS_SQL =  """
+        const val READ_UNPROCESSED_ORDER_OPERATIONS_FOR_UPDATE_SQL =  """
                 SELECT id, order_id, order_statuses.merchant_id, status_change_time,
                     operation_status AS order_operation_status, record_read_count, order_status
                 FROM scheduler_quartz.order_statuses
                 WHERE
                     operation_status IN ('READY_FOR_PROCESSING', 'FOR_RETRY')
-                    AND merchant_id IN (:merchantIds)
+                    AND merchant_id NOT IN (:excludedMerchantIds)
                     AND order_statuses.order_status IN (:orderStatuses)
                 LIMIT :maxPageSize
                 FOR UPDATE SKIP LOCKED
         """
 
-        // Todo: independently execute update after Kafka + limit max page size
         @Language("PostgreSQL")
         const val INCREASE_UNPROCESSED_ORDER_OPERATION_READ_COUNT_SQL = """
             UPDATE scheduler_quartz.order_statuses
@@ -54,11 +54,11 @@ class WildFruitOperationOnOrderRepository(
     @Transactional(propagation = Propagation.REQUIRES_NEW)
     fun readUnprocessedWithReadCountIncrement(
         maxPageSize: Long,
-        orderJobData: DedicatedOrderJobData
+        orderJobData: CommonOrderJobData
     ): List<OperationOnOrder> {
-        val updatedOrderStatuses = jdbcClient.sql(READ_UNPROCESSED_ORDER_OPERATIONS_SQL)
+        val updatedOrderStatuses = jdbcClient.sql(READ_UNPROCESSED_ORDER_OPERATIONS_FOR_UPDATE_SQL)
             .param("maxPageSize", maxPageSize)
-            .param("merchantIds", orderJobData.merchantIds)
+            .param("excludedMerchantIds", orderJobData.excludedMerchantIds)
             .param("orderStatuses", orderJobData.orderStatuses.asSequence().map { it.value }.toSet())
             .query(OperationOnOrder::class.java)
             .list()
@@ -74,19 +74,12 @@ class WildFruitOperationOnOrderRepository(
             .list()
     }
 
-    fun markOrderOperationsFailed(orderIds: Set<Long>) = markOrderOperationsAs(orderIds, OrderOperationStatus.ERROR)
-
     fun markOrderOperationsAsProcessed(orderIds: Set<Long>)
-        = markOrderOperationsAs(orderIds, OrderOperationStatus.SENT_TO_NOTIFIER)
-
-    fun markOrderOperationsAs(orderIds: Set<Long>, status: OrderOperationStatus)
         = if (orderIds.isNotEmpty()) {
             jdbcClient.sql(CHANGE_ORDER_OPERATION_STATUSES_SQL)
-                .param("orderOperationStatus", status, Types.VARCHAR)
-                .param("ids", orderIds)
+                .param("orderOperationStatus", OrderOperationStatus.SENT_TO_NOTIFIER, Types.VARCHAR)
                 .param("statusChangeTime", LocalDateTime.now())
+                .param("ids", orderIds)
                 .update()
     } else 0
-
-
 }
