+++ scheduler-quartz/src/main/kotlin/io/scheduler/comparison/quartz/repositories/pagination/WildFruitOperationOnOrderRepository.kt
@@ -1,9 +1,10 @@
-package io.scheduler.comparison.quartz.repositories
+package io.scheduler.comparison.quartz.repositories.pagination
 
 import io.scheduler.comparison.quartz.domain.OperationOnOrder
 import io.scheduler.comparison.quartz.domain.OrderOperationStatus
-import io.scheduler.comparison.quartz.jobs.state.CommonOrderJobData
+import io.scheduler.comparison.quartz.jobs.state.DedicatedOrderJobData
 import org.intellij.lang.annotations.Language
+import org.springframework.context.annotation.Profile
 import org.springframework.jdbc.core.simple.JdbcClient
 import org.springframework.stereotype.Repository
 import org.springframework.transaction.annotation.Propagation
@@ -11,21 +12,21 @@ import org.springframework.transaction.annotation.Transactional
 import java.sql.Types
 import java.time.LocalDateTime
 
-// Todo: queryForStream with Stream item processing (pageSize == fetchSize), maxCount == LIMIT.
 @Repository
-class CommonOperationOnOrderRepository(
+@Profile("pagination")
+class WildFruitOperationOnOrderRepository(
     private val jdbcClient: JdbcClient
 ) {
 
     private companion object {
         @Language("PostgreSQL")
-        const val READ_UNPROCESSED_ORDER_OPERATIONS_FOR_UPDATE_SQL =  """
+        const val READ_UNPROCESSED_ORDER_OPERATIONS_SQL =  """
                 SELECT id, order_id, order_statuses.merchant_id, status_change_time,
                     operation_status AS order_operation_status, record_read_count, order_status
                 FROM scheduler_quartz.order_statuses
                 WHERE
                     operation_status IN ('READY_FOR_PROCESSING', 'FOR_RETRY')
-                    AND merchant_id NOT IN (:excludedMerchantIds)
+                    AND merchant_id IN (:merchantIds)
                     AND order_statuses.order_status IN (:orderStatuses)
                 LIMIT :maxPageSize
                 FOR UPDATE SKIP LOCKED
@@ -54,11 +55,11 @@ class CommonOperationOnOrderRepository(
     @Transactional(propagation = Propagation.REQUIRES_NEW)
     fun readUnprocessedWithReadCountIncrement(
         maxPageSize: Long,
-        orderJobData: CommonOrderJobData
+        orderJobData: DedicatedOrderJobData
     ): List<OperationOnOrder> {
-        val updatedOrderStatuses = jdbcClient.sql(READ_UNPROCESSED_ORDER_OPERATIONS_FOR_UPDATE_SQL)
+        val updatedOrderStatuses = jdbcClient.sql(READ_UNPROCESSED_ORDER_OPERATIONS_SQL)
             .param("maxPageSize", maxPageSize)
-            .param("excludedMerchantIds", orderJobData.excludedMerchantIds)
+            .param("merchantIds", orderJobData.merchantIds)
             .param("orderStatuses", orderJobData.orderStatuses.asSequence().map { it.value }.toSet())
             .query(OperationOnOrder::class.java)
             .list()
@@ -74,12 +75,19 @@ class CommonOperationOnOrderRepository(
             .list()
     }
 
+    fun markOrderOperationsFailed(orderIds: Set<Long>) = markOrderOperationsAs(orderIds, OrderOperationStatus.ERROR)
+
     fun markOrderOperationsAsProcessed(orderIds: Set<Long>)
+        = markOrderOperationsAs(orderIds, OrderOperationStatus.SENT_TO_NOTIFIER)
+
+    fun markOrderOperationsAs(orderIds: Set<Long>, status: OrderOperationStatus)
         = if (orderIds.isNotEmpty()) {
             jdbcClient.sql(CHANGE_ORDER_OPERATION_STATUSES_SQL)
-                .param("orderOperationStatus", OrderOperationStatus.SENT_TO_NOTIFIER, Types.VARCHAR)
-                .param("statusChangeTime", LocalDateTime.now())
+                .param("orderOperationStatus", status, Types.VARCHAR)
                 .param("ids", orderIds)
+                .param("statusChangeTime", LocalDateTime.now())
                 .update()
     } else 0
+
+
 }
